Factor out table heading
